build_all_repos: Find perl targets after compilation
The previous code assumed the location of the Perl lib directories for each repo, but as POMs represent a hierarchical
structure of dependencies this was not always valid, most notably in the case of the configuration-modules-* repos.
This only became an issue when AII became able to use components directly and test cases started making use of this.
The fix also required the special cases for the maven-tools repository to be separated.